requirements_dev.txt
python-dotenv
numpy==1.26.4
pandas==2.2.1
mlflow==2.12.1
scikit-learn==1.5.0
hyperopt==0.2.7
xgboost==2.1.0
boto3==1.34.144
requests==2.32.3